+++ Code/Simple/PreallocatedPrecomputed/Tree.cpp
@@ -22,7 +22,7 @@ Tree::Tree(vector<uint>* input, uint amount, uint alphabetSize, uint blockSize)
 
     bitmap = new bitmap_t(maxBitmapSize);
 
-    uint bitmapOffset = 0;
+    ulong bitmapOffset = 0;
     root = Node(input, (uint)0, alphabetSize-1, nullptr, bitmap, bitmapOffset, blockRanks, blockSize);
 
 //    bitmap->resize(bitmapOffset);
